backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
chromadb==0.5.5
openai==1.51.0
anthropic==0.34.2
gitpython==3.1.43
pydantic==2.9.2
python-dotenv==1.0.1
